# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Just VPN 后端接口文档",
        "title": "Just VPN API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/lucky_wheel_get_reward": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户领取当天的中奖奖品",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "活动"
                ],
                "summary": "领取幸运转盘奖励",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/lucky_wheel_get_status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回当前用户当天是否已参与过幸运转盘，以及新用户开关和通用开关的状态。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "活动"
                ],
                "summary": "获取幸运转盘状态",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/controller.LuckyWheelStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/lucky_wheel_play": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "通用开关lucky_wheel.general_enabled为on，或新用户开关lucky_wheel.new_user_enabled为on且user.type=1，满足任一条件即可参与，不限制注册时长。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "活动"
                ],
                "summary": "随机获取幸运转盘奖品",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/controller.LuckyWheelPlayResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/lucky_wheel_winners": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "随机生成30条中奖记录，用户ID为随机数字前后拼接的字符串，长度不固定。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "活动"
                ],
                "summary": "获取用户中奖记录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/controller.LuckyWheelWinnerResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.LuckyWheelPlayResponse": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string",
                    "example": "免费会员"
                },
                "level": {
                    "type": "integer",
                    "example": 1
                },
                "remark": {
                    "type": "string",
                    "example": "高速会员专属权益"
                },
                "title": {
                    "type": "string",
                    "example": "10分钟免费会员"
                }
            }
        },
        "controller.LuckyWheelStatusResponse": {
            "type": "object",
            "properties": {
                "general_enabled": {
                    "type": "string",
                    "example": "off"
                },
                "news_enabled": {
                    "type": "string",
                    "example": "off"
                },
                "today_played": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "controller.LuckyWheelWinnerResponse": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string",
                    "example": "免费会员"
                },
                "remark": {
                    "type": "string",
                    "example": "高速会员专属权益"
                },
                "title": {
                    "type": "string",
                    "example": "10分钟"
                },
                "user_id": {
                    "type": "string",
                    "example": "123...4"
                }
            }
        },
...
        "controller.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                },
                "result": {}
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}